float values finance calculator

+++ finance_calculator.py
@@ -1,6 +1,6 @@
 monthly_income = input("Enter your monthly income: ")
 monthly_expenses =  input("Enter your total monthly expenses: ")
-monthly_savings = int(monthly_income) - int(monthly_expenses)
+monthly_savings = float(monthly_income) - float(monthly_expenses)
 
 projected_savings = monthly_savings * 12 * 1.05
